Segmentation fault & Logfiles
Unified the schema which is use to store the log files. Now log files
are saved for each processing step beginning with the subject-ID
followed by the name of the step itself. E.g. sub1_preprocessing.log
Furthermore changed the way the tracking script is called since reading
out parameters of the locally stored file seed count.txt caused
segmentation faults on the stampede cluster, probably due to memory
locks when processing tracking in parallel.
The parameters stored in this txt-File are now directly passed to the
tracking-Batch-Script to avoid this issue.